fsl-eula-unpack.bbclass: Enhance EULA model
Each new version of the EULA is meant to be a click
through that includes any existing packages with an
older EULA. The latest EULA is also copied in the
layer.

Capture this model in the class by listing the
known EULAs and checking that any package
has a known EULA. The new model also allows
extensibility, so new EULAs can be introduced
separately.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>
(cherry picked from commit 9418435542ce76a2346f1a4aa2cbbb429a5d3543)